test that event data has correct fields in bar, scatter, histogram
